// src/interfaces/users.ts
export interface users {
    _id?: string
    name: {
        first: string;
        middle?: string;
        last: string;
    };
    phone: string;
    email: string;
    password: string;
    image: {
        url: string;
        alt: string;
    };
    address: {
        state?: string;
        country: string;
        city: string;
        street: string;
        houseNumber: number;
        zip: number;
    };
    isBusiness: boolean
    isAdmin?: boolean;
}